fix: folder organization
